Move macro helpers into helpers submodule
This is in preperation for new marcos which are for example used
for netlink related functions, and therefore having one big macro
file is not practicable anymore.
